Split out RestApiServerError into file
This commit moves the `RestApiServerError` type into its own file to
keep the errors/mod.rs file clean and conciser.
